LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/hit-test-in-vertical-rl-expected.txt


FAIL Hit-testing in vertical-rl writing-mode assert_equals: expected 5 but got -155
